Generalize the R installation
The script for installing R should now apply to other versions that the
user might choose to use, although it currently focuses on the more
recent versions, due to the original purpose for which the script was
developed.

We now check for existing installation of R, RStudio and Rtools. Still
much work to be done, but the general idea has been set.